% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BICselect.R
\name{BICselect}
\alias{BICselect}
\title{Convolutive regression best model selection BIC}
\usage{
BICselect(
  df,
  formula.resp,
  idx.pred,
  dist1,
  dist2,
  quiet = TRUE,
  formulas = "main",
  parallel = FALSE,
  interactions = FALSE,
  debug = FALSE,
  ...
)
}
\arguments{
\item{df}{dataset}

\item{formula.resp}{response formula}

\item{idx.pred}{indices of predictors}

\item{dist1}{chain of character to idnetify the distribution of variable1}

\item{dist2}{chain of character to idnetify the distribution of variable2}

\item{quiet}{logical. Displaying the progress.}

\item{formulas}{"main" or "all". All includes the variances.}

\item{parallel}{logical. Should the function run on parallel threads.}

\item{interactions}{logical. Should the formulas include interactions.}

\item{debug}{logical. Working option to debug. Dvp.}

\item{...}{Additional arguments fed into convreg().}
}
\description{
Convolutive regression best model selection BIC
}
\examples{
set.seed(123)
e=rnorm(n=50,mean=0,sd=0.05)
x = rnorm(50,0,0.5)
k= rnbinom(50,exp(x),0.5)
y= data.frame(obs= k + e +x, f1 = x, f2 = k, f3 = 2*rnorm(n=50,mean=0,sd=0.05))
formula.resp = ~obs
df = y
idx.pred = 2
dist1 = "Nbinom"
dist2 = "Gauss"
res= BICselect(df = df, formula.resp = formula.resp, idx.pred = idx.pred, dist1 = dist1, dist2 = dist2)
head(res,10)
}
\keyword{BIC}
