% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfplot.R
\name{cdfplot}
\alias{cdfplot}
\title{Convolutive regression function cdf plot}
\usage{
cdfplot(obj, delta = 0.5, lwd.pt = 2, cex.pt = 1, cex.text = 1, ...)
}
\arguments{
\item{obj}{convreg object}

\item{delta}{discretization parameter}

\item{lwd.pt}{line width for observations}

\item{cex.pt}{point size for observations}

\item{cex.text}{legend size.}

\item{...}{additional plotting parameters}
}
\description{
Convolutive regression function cdf plot
}
\examples{
set.seed(123)
e=rnorm(n=150,mean=0,sd=0.05)
x = rbinom(150,1,0.5)
k=5 + x
y= data.frame(obs=k + e , f1 = as.factor(x))
res=convreg( ~obs,
          formula.mu1 =~ 1,
          formula.mu2 =~ f1,
          data=y)
cdfplot(res)
}
\keyword{convreg}
