% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convReg-package.R
\docType{package}
\name{convReg-package}
\alias{convReg}
\alias{convReg-package}
\title{convReg : Package for convolutive mixture models.}
\description{
convReg allows you to infer convolutive mixture models.
}
\author{
\strong{Maintainer}: Adrien Ickowicz \email{adrien.ickowicz@csiro.au}

}
