% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convreg.R
\name{convreg}
\alias{convreg}
\title{Convolutive regression function}
\usage{
convreg(
  formula.resp,
  formula.mu1 = ~1,
  formula.sigma1 = ~1,
  formula.mu2 = ~1,
  formula.sigma2 = ~1,
  data,
  dist1 = "Nbinom",
  dist2 = "Gauss",
  fixed = NULL,
  na.action,
  method = "mle",
  debug = FALSE,
  quiet = TRUE,
  weights = NULL,
  scale = FALSE,
  scaleInit = 1,
  ...
)
}
\arguments{
\item{formula.resp}{response}

\item{formula.mu1}{formula for variable1 expectation}

\item{formula.sigma1}{formula for variable1 variance}

\item{formula.mu2}{formula for variable2 expectation}

\item{formula.sigma2}{formula for variable2 variance}

\item{data}{dataset}

\item{dist1}{chain of character to idnetify the distribution of variable1}

\item{dist2}{chain of character to idnetify the distribution of variable2}

\item{fixed}{set the parameters that are fixed to some specified values}

\item{na.action}{action for NA values}

\item{method}{estimation method}

\item{debug}{getting into the function}

\item{quiet}{display comments.}

\item{weights}{weights.}

\item{scale}{Should the observations be scales}

\item{scaleInit}{Initial value of the scaling.}

\item{...}{additional parameters}
}
\description{
Convolutive regression function
}
\examples{
set.seed(123)
e=0.75+rnorm(n=500,mean=0,sd=0.25)
x1 = rnorm(500,0,0.5)
x2 = runif(500,-2,1)
k= rnbinom(500,mu=exp(x2),size=1.5)
y= data.frame(obs=(k + e) , f1 = x1, f2 = x2)
par(mfrow = c(1,1), mar = c(5,4,4,2))
hist(y$obs, breaks = seq(min(y$obs)-1.5, max(y$obs)+3, 3), xlim = c(0,14*24))
res.reg.em =convreg( ~obs,
          formula.mu1 =~ f2,
          formula.mu2 =~ f1,
          data=y,
          dist1 = "Nbinom",
          method = "mle")
res.reg.em
summary(res.reg.em)
distplot(res.reg.em, delta=3, xlim=c(0,200))
}
\keyword{convreg}
