% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dBinomGauss}
\alias{dBinomGauss}
\title{Binomial Gaussian convolution}
\usage{
dBinomGauss(x, prob, size, mu, sigma, log = FALSE)
}
\arguments{
\item{x}{observations}

\item{prob}{Binomial probability}

\item{size}{Binomial size}

\item{mu}{Gaussian mean}

\item{sigma}{Gaussian variance}

\item{log}{Logical. Should the value be on log scale.}
}
\description{
Binomial Gaussian convolution
}
\keyword{Binomial}
\keyword{Gaussian}
\keyword{convolution}
