% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dBinomLnorm}
\alias{dBinomLnorm}
\title{Binomial Lognormal convolution}
\usage{
dBinomLnorm(x, prob, size, mu, sigma, log = FALSE)
}
\arguments{
\item{x}{observations}

\item{prob}{Binomial probability}

\item{size}{Binomial size}

\item{mu}{Log Gaussian mean}

\item{sigma}{Log Gaussian variance}

\item{log}{Logical. Should the value be on log scale.}
}
\description{
Binomial Lognormal convolution
}
\keyword{Binomial}
\keyword{Gaussian}
\keyword{Log}
\keyword{convolution}
