% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dcomp}
\alias{dcomp}
\title{Conway Maxwell Poisson distribution}
\usage{
dcomp(y, lam, nu, sumTo = 50L, logP = FALSE)
}
\arguments{
\item{y}{observations}

\item{lam}{Mean Poisson paramter}

\item{nu}{Variance parameter}

\item{sumTo}{integer value for calculating the distribution. Default to 50.}

\item{logP}{Logical. Should the value be on log scale.}
}
\description{
Conway Maxwell Poisson distribution
}
\keyword{Conway}
\keyword{Maxwell}
\keyword{Poisson}
