% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dCoMPoissonGauss2}
\alias{dCoMPoissonGauss2}
\title{Conway Maxwell Poisson Gaussian convolution}
\usage{
dCoMPoissonGauss2(x, mu, size, mug, sigmag, log = FALSE)
}
\arguments{
\item{x}{observations}

\item{mu}{CoMP mean}

\item{size}{CoMP size}

\item{mug}{Gaussian mean}

\item{sigmag}{Gaussian variance}

\item{log}{Logical. Should the value be on log scale.}
}
\description{
Conway Maxwell Poisson Gaussian convolution
}
\keyword{CoMP}
\keyword{Gaussian}
\keyword{convolution}
