% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DebTrivedi}
\alias{DebTrivedi}
\title{Patient medication data}
\format{
A data frame with columns:
\describe{
\item{ofp}{}
\item{ofnp}{}
\item{opp}{}
\item{opnp}{}
\item{emer}{}
\item{hosp}{}
\item{health}{}
\item{numchron}{}
\item{adldiff}{}
\item{region}{}
\item{age}{}
\item{black}{}
\item{gender}{}
\item{married}{}
\item{school}{}
\item{faminc}{}
\item{employed}{}
\item{privins}{}
\item{medicaid}{}
}
}
\usage{
DebTrivedi
}
\description{
Patient medication data
}
\examples{
\dontrun{
 DebTrivedi
}
}
\keyword{datasets}
