% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dHP}
\alias{dHP}
\title{Hurdle Poisson distribution}
\usage{
dHP(y, lam, pi, logP = FALSE)
}
\arguments{
\item{y}{observations}

\item{lam}{Mean Poisson paramter}

\item{pi}{Hurdle probability}

\item{logP}{Logical. Should the value be on log scale.}
}
\description{
Hurdle Poisson distribution
}
\keyword{Poisson}
\keyword{hurdle}
