% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dHPGauss}
\alias{dHPGauss}
\title{Hurdle Poisson Gaussian convolution}
\usage{
dHPGauss(x, lam, pi, mug, sigmag, log = FALSE)
}
\arguments{
\item{x}{observations}

\item{lam}{Hurdle mean}

\item{pi}{Hurdle probability}

\item{mug}{Gaussian mean}

\item{sigmag}{Gaussian variance}

\item{log}{Logical. Should the value be on log scale.}
}
\description{
Hurdle Poisson Gaussian convolution
}
\keyword{Gaussian}
\keyword{ZIP}
\keyword{convolution}
