% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{dMultinomGamma}
\alias{dMultinomGamma}
\title{Mulinomial and Gamma convolution}
\usage{
dMultinomGamma(x, mu, size, mug, sigmag, log = F)
}
\arguments{
\item{x}{value}

\item{mu}{parameter1-1}

\item{size}{parameter1-2}

\item{mug}{parameter2-1}

\item{sigmag}{parameter2-2}

\item{log}{output on log-scale?}
}
\description{
Mulinomial and Gamma convolution
}
\keyword{convolution}
