% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dNbinomGauss}
\alias{dNbinomGauss}
\title{Negative Binomial Gaussian convolution}
\usage{
dNbinomGauss(x, mu, size, mug, sigmag, log = FALSE)
}
\arguments{
\item{x}{observations}

\item{mu}{Negative Binomial mean}

\item{size}{Negative Binomial size}

\item{mug}{Gaussian mean}

\item{sigmag}{Gaussian variance}

\item{log}{Logical. Should the value be on log scale.}
}
\description{
Negative Binomial Gaussian convolution
}
\keyword{Binomial}
\keyword{Gaussian}
\keyword{Negative}
\keyword{convolution}
