% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dNbinomLnorm}
\alias{dNbinomLnorm}
\title{Negative Binomial Log Gaussian convolution}
\usage{
dNbinomLnorm(x, mu, size, mug, sigmag, log = FALSE)
}
\arguments{
\item{x}{observations}

\item{mu}{Negative Binomial mean}

\item{size}{Negative Binomial size}

\item{mug}{Log Gaussian mean}

\item{sigmag}{Log Gaussian variance}

\item{log}{Logical. Should the value be on log scale.}
}
\description{
Negative Binomial Log Gaussian convolution
}
\keyword{Binomial}
\keyword{Gaussian}
\keyword{Log}
\keyword{Negative}
\keyword{convolution}
