% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dPoisGauss}
\alias{dPoisGauss}
\title{Poisson Gaussian convolution}
\usage{
dPoisGauss(x, lam, mug, sigmag, log = FALSE)
}
\arguments{
\item{x}{observations}

\item{lam}{Poisson mean}

\item{mug}{Gaussian mean}

\item{sigmag}{Gaussian variance}

\item{log}{Logical. Should the value be on log scale.}
}
\description{
Poisson Gaussian convolution
}
\keyword{Gaussian}
\keyword{Poisson}
\keyword{convolution}
