% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dPoisLnorm}
\alias{dPoisLnorm}
\title{Poisson Log Gaussian convolution}
\usage{
dPoisLnorm(x, lam, mu, sigma, log = FALSE)
}
\arguments{
\item{x}{observations}

\item{lam}{Poisson mean}

\item{mu}{Log Gaussian mean}

\item{sigma}{Log Gaussian variance}

\item{log}{Logical. Should the value be on log scale.}
}
\description{
Poisson Log Gaussian convolution
}
\keyword{Gaussian}
\keyword{Log}
\keyword{Poisson}
\keyword{convolution}
