% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dtpois}
\alias{dtpois}
\title{Zero-truncated Poisson distribution}
\usage{
dtpois(y, lam, logP = FALSE)
}
\arguments{
\item{y}{observations}

\item{lam}{Mean Poisson paramter}

\item{logP}{Logical. Should the value be on log scale.}
}
\description{
Zero-truncated Poisson distribution
}
\keyword{Poisson}
\keyword{truncated}
\keyword{zero}
