% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dZIP}
\alias{dZIP}
\title{Zero-inflated Poisson distribution}
\usage{
dZIP(y, lam, pi, sumTo = 100L, logP = FALSE)
}
\arguments{
\item{y}{observations}

\item{lam}{Mean Poisson paramter}

\item{pi}{Inflation probability}

\item{sumTo}{integer value for calculating the distribution. Default to 100.}

\item{logP}{Logical. Should the value be on log scale.}
}
\description{
Zero-inflated Poisson distribution
}
\keyword{Poisson}
\keyword{inflated}
\keyword{zero}
