% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dZIPGauss}
\alias{dZIPGauss}
\title{Zero inflated Poisson Gaussian convolution}
\usage{
dZIPGauss(x, lam, pi, mug, sigmag, log = FALSE)
}
\arguments{
\item{x}{observations}

\item{lam}{ZIP mean}

\item{pi}{ZIP probability}

\item{mug}{Gaussian mean}

\item{sigmag}{Gaussian variance}

\item{log}{Logical. Should the value be on log scale.}
}
\description{
Zero inflated Poisson Gaussian convolution
}
\keyword{Gaussian}
\keyword{ZIP}
\keyword{convolution}
