% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmComp.R
\name{glmComp}
\alias{glmComp}
\title{Conway-Maxwell-Poisson GLM}
\usage{
glmComp(
  lamFormula,
  nuFormula = NULL,
  data,
  weights = NULL,
  lamStart = NULL,
  nuStart = NULL,
  sumTo = 100L,
  method = "BFGS",
  ...
)
}
\arguments{
\item{lamFormula}{formula for lambda}

\item{nuFormula}{formula for nu}

\item{data}{dataset}

\item{weights}{weights}

\item{lamStart}{initial values for lambda}

\item{nuStart}{initial values for nu}

\item{sumTo}{integer to sum to to calculate the normalization constant}

\item{method}{optimisation method}

\item{...}{additional parameters}
}
\description{
Conway-Maxwell-Poisson GLM function
}
\examples{
# Not run
}
\keyword{comp}
