% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{marginalReg}
\alias{marginalReg}
\title{Calculate marginal regression}
\usage{
marginalReg(obj, name.reg, bp, ...)
}
\arguments{
\item{obj}{convreg object}

\item{name.reg}{variable on which calculate the regression}

\item{bp}{boolean. If factor.}

\item{...}{additional parameters}
}
\description{
Calculate marginal regression
}
\keyword{convolution}
