% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{meanVarianceConvreg}
\alias{meanVarianceConvreg}
\title{Reports total mean and total variance of convreg distribution}
\usage{
meanVarianceConvreg(mu1, sigma1, mu2, sigma2, dist1, dist2)
}
\arguments{
\item{mu1}{value for mu1}

\item{sigma1}{value for sigma1}

\item{mu2}{value for mu2}

\item{sigma2}{value for sigma2}

\item{dist1}{choice of distribution 1}

\item{dist2}{choice of distribution 2}
}
\description{
Reports total mean and total variance of convreg distribution
}
