% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatientAdmission-data.R
\docType{data}
\name{PatientAdmission}
\alias{PatientAdmission}
\title{Patient Admission data}
\format{
An object of class
}
\usage{
data(PatientAdmission)
}
\description{
Data from a set of hospitals in Victoria (Australia)
}
\examples{
data(PatientAdmission)
}
\references{
Ickowicz et al. (2016) Statistics in Medicine 195:1077-1086
}
\keyword{datasets}
