% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{pBinomGamma}
\alias{pBinomGamma}
\title{Binomial and Gamma convolution}
\usage{
pBinomGamma(q, prob, size, mug, sigmag)
}
\arguments{
\item{q}{quantiles}

\item{prob}{parameter1-1}

\item{size}{parameter1-2}

\item{mug}{parameter2-1}

\item{sigmag}{parameter2-2}
}
\description{
Binomial and Gamma convolution
}
\keyword{convolution}
