% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{pBinomGauss}
\alias{pBinomGauss}
\title{Binomial and Gaussian convolution}
\usage{
pBinomGauss(q, prob, size, mug, sigmag)
}
\arguments{
\item{q}{quantiles}

\item{prob}{parameter1-1}

\item{size}{parameter1-2}

\item{mug}{parameter2-1}

\item{sigmag}{parameter2-2}
}
\description{
Binomial and Gaussian convolution
}
\keyword{convolution}
