% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{pBinomLnorm}
\alias{pBinomLnorm}
\title{Binomial and Log-Gaussian convolution}
\usage{
pBinomLnorm(q, prob, size, mug, sigmag)
}
\arguments{
\item{q}{quantiles}

\item{prob}{parameter1-1}

\item{size}{parameter1-2}

\item{mug}{parameter2-1}

\item{sigmag}{parameter2-2}
}
\description{
Binomial and Log-Gaussian convolution
}
\keyword{convolution}
