% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{pCoMPoissonGauss}
\alias{pCoMPoissonGauss}
\title{CoMP and Gaussian convolution}
\usage{
pCoMPoissonGauss(q, mu, size, mug, sigmag)
}
\arguments{
\item{q}{quantiles}

\item{mu}{parameter1-1}

\item{size}{parameter1-2}

\item{mug}{parameter2-1}

\item{sigmag}{parameter2-2}
}
\description{
CoMP and Gaussian convolution
}
\keyword{convolution}
