% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTestdist.R
\name{plotTestdist}
\alias{plotTestdist}
\title{Convolutive regression function testing plot}
\usage{
plotTestdist(testdist)
}
\arguments{
\item{testdist}{testdist object}
}
\description{
Convolutive regression function testing plot
}
\examples{
set.seed(123)
e=rlnorm(n=150,mean=0,sd=0.05)
x = rnbinom(150,1,0.5)
k=5 + x
y= data.frame(obs=k + e , f1 = as.factor(x))
resTest=testdist( y$obs,
          formula.mu1 =~ 1,
          formula.mu2 =~ f1,
          data=y)
plotTestdist(resTest)
}
\keyword{convreg}
