% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{pPoisGamma}
\alias{pPoisGamma}
\title{Poisson and Gamma convolution}
\usage{
pPoisGamma(q, mu, mug, sigmag)
}
\arguments{
\item{q}{quantiles}

\item{mu}{parameter1-1}

\item{mug}{parameter2-1}

\item{sigmag}{parameter2-2}
}
\description{
Poisson and Gamma convolution
}
\keyword{convolution}
