% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{pPoisGauss}
\alias{pPoisGauss}
\title{Poisson and Gaussian convolution}
\usage{
pPoisGauss(q, mu, mug, sigmag)
}
\arguments{
\item{q}{quantiles}

\item{mu}{parameter1-1}

\item{mug}{parameter2-1}

\item{sigmag}{parameter2-2}
}
\description{
Poisson and Gaussian convolution
}
\keyword{convolution}
