% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{pPoisLnorm}
\alias{pPoisLnorm}
\title{Poisson and Log-Gaussian convolution}
\usage{
pPoisLnorm(q, mu, mug, sigmag)
}
\arguments{
\item{q}{quantiles}

\item{mu}{parameter1-1}

\item{mug}{parameter2-1}

\item{sigmag}{parameter2-2}
}
\description{
Poisson and Log-Gaussian convolution
}
\keyword{convolution}
