% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{pZIPGauss}
\alias{pZIPGauss}
\title{ZIP and Gaussian convolution}
\usage{
pZIPGauss(q, mu, size, mug, sigmag)
}
\arguments{
\item{q}{quantiles}

\item{mu}{parameter1-1}

\item{size}{parameter1-2}

\item{mug}{parameter2-1}

\item{sigmag}{parameter2-2}
}
\description{
ZIP and Gaussian convolution
}
\keyword{convolution}
