% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{qNbinomGauss}
\alias{qNbinomGauss}
\title{Negative binomial and Gaussian convolution}
\usage{
qNbinomGauss(p, mu, size, mug, sigmag)
}
\arguments{
\item{p}{probabilities}

\item{mu}{parameter1-1}

\item{size}{parameter1-2}

\item{mug}{parameter2-1}

\item{sigmag}{parameter2-2}
}
\description{
Negative binomial and Gaussian convolution
}
\keyword{convolution}
