% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{qPoisGamma}
\alias{qPoisGamma}
\title{Poisson and Gamma convolution}
\usage{
qPoisGamma(p, mu, size, mug, sigmag)
}
\arguments{
\item{p}{probabilities}

\item{mu}{parameter1-1}

\item{size}{parameter1-2}

\item{mug}{parameter2-1}

\item{sigmag}{parameter2-2}
}
\description{
Poisson and Gamma convolution
}
\keyword{convolution}
