% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_ci.R
\name{qq_ci}
\alias{qq_ci}
\title{QQ plot}
\usage{
qq_ci(
  x,
  distribution = "norm",
  ...,
  line.estimate = NULL,
  conf = 0.95,
  labels = names(x)
)
}
\arguments{
\item{x}{a sample}

\item{distribution}{a distribution name}

\item{...}{additional plotting parameters}

\item{line.estimate}{line estimate}

\item{conf}{bounds of the confidence region}

\item{labels}{names}
}
\description{
QQ plot for unusual distributions
}
\keyword{QQ}
\keyword{plot}
