% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{quantRes}
\alias{quantRes}
\title{Calculate quantile residuals}
\usage{
quantRes(obj, ...)
}
\arguments{
\item{obj}{convreg object}

\item{...}{additional parameters}
}
\description{
Calculate quantile residuals
}
\keyword{convolution}
