% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{rCoMPoissonLnorm}
\alias{rCoMPoissonLnorm}
\title{CoMP and Log-Gaussian convolution}
\usage{
rCoMPoissonLnorm(N, mu, size, mug, sigmag)
}
\arguments{
\item{N}{sample size}

\item{mu}{parameter1-1}

\item{size}{parameter1-2}

\item{mug}{parameter2-1}

\item{sigmag}{parameter2-2}
}
\description{
CoMP and Log-Gaussian convolution
}
\keyword{convolution}
