% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regplot.R
\name{regplot}
\alias{regplot}
\title{Convolutive regression function regression plot}
\usage{
regplot(obj, delta = 0.5, disp = "all", cex.pt = 0.5, ...)
}
\arguments{
\item{obj}{convreg object}

\item{delta}{discretization parameter}

\item{disp}{choice of predictors to plot}

\item{cex.pt}{point size for observations}

\item{...}{additional plotting parameters}
}
\description{
Convolutive regression function regression plot
}
\examples{
set.seed(123)
x1 = rnorm(500,0.75,0.25)
x2 = rnorm(500,-2,1)
k= rnbinom(500,mu=exp(x2),size=0.5)
e= x1 + rnorm(n=500,mean=0,sd=0.25)
y= data.frame(obs=k + e , f1 = x1, f2 = x2)
res=convreg( ~obs,
          formula.mu1 =~ f2,
          formula.mu2 =~ f1,
          data=y)
regplot(res, cex.pt = 0.25)
}
\keyword{convreg}
