% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{rHPGauss}
\alias{rHPGauss}
\title{HP and Gaussian convolution}
\usage{
rHPGauss(N, mu, pi, mug, sigmag)
}
\arguments{
\item{N}{sample size}

\item{mu}{parameter1-1}

\item{pi}{parameter1-2}

\item{mug}{parameter2-1}

\item{sigmag}{parameter2-2}
}
\description{
HP and Gaussian convolution
}
\keyword{convolution}
