% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{rtpois}
\alias{rtpois}
\title{Zero-truncated Poisson distribution}
\usage{
rtpois(N, lambda)
}
\arguments{
\item{N}{sample size}

\item{lambda}{mean Poisson parameter}
}
\description{
Zero-truncated Poisson distribution
}
\keyword{Poisson}
\keyword{truncated}
