% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdist.R
\name{testdist}
\alias{testdist}
\title{Convolutive regression function testing}
\usage{
testdist(y.obs, dist1 = "Nbinom", dist2 = "Gauss", ...)
}
\arguments{
\item{y.obs}{observations vector}

\item{dist1}{chain of character to idnetify the distribution of variable1}

\item{dist2}{chain of character to idnetify the distribution of variable2}

\item{...}{additional parameters for convreg().}
}
\description{
Convolutive regression function testing
}
\examples{
set.seed(123)
e=rnorm(n=150,mean=0,sd=0.05)
x = rbinom(150,1,0.5)
k=5 + x
y= data.frame(obs=k + e , f1 = as.factor(x))
resTest=testdist(y$obs)
}
\keyword{convreg}
\keyword{test}
