% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DebTrivedi-data.R
\docType{data}
\name{DebTrivedi}
\alias{DebTrivedi}
\title{Patient medication data}
\format{
An object of class
}
\usage{
data(DebTrivedi)
}
\description{
Data from a set of patients
}
\examples{
data(DebTrivedi)
}
\references{
Ickowicz et al. (2016) Statistics in Medicine 195:1077-1086
}
\keyword{datasets}
