% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PatientAdmission}
\alias{PatientAdmission}
\title{Patient Admission data}
\format{
A data frame with columns:
\describe{
\item{LOS.total}{Lenght of stay}
\item{Age}{Age of patient}
\item{NumberEpisodes}{Number of hospital episodes}
\item{RealAdmissionDate}{Date of admission}
\item{RealSeparationDate}{Date of discharge}
\item{GenderName}{Gender of patient}
\item{Postcode}{Postcode of patient}
\item{MaritalStatusNameCorr}{Marital status of patient}
\item{DiseaseTypesCorr}{Disease type - 9 levels}
}
}
\source{
Data from a set of hospitals in Victoria (Australia)
}
\usage{
PatientAdmission
}
\description{
Patient Admission data
}
\examples{
\dontrun{
 PatientAdmission
}
}
\keyword{datasets}
